app/requirements.txt
alembic==1.0.8
amqp==2.4.2
b2blaze==0.2.1
bcrypt==3.1.6
billiard==3.5.0.5
blinker==1.4
boto3==1.9.108
botocore==1.12.108
celery==4.2.1
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudinary==1.15.0
decorator==4.3.2
docutils==0.14
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Redis==0.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
funcy==1.11
idna==2.7
infinity==1.4
intervals==0.8.1
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.4
kombu==4.4.0
Mako==1.0.7
markdown2==2.3.7
MarkupSafe==1.1.1
mock==2.0.0
pbr==5.1.3
Pillow==5.4.1
psycopg2-binary==2.7.7
pycparser==2.19
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
python-slugify==3.0.0
pytz==2018.9
redis==3.2.0
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
SQLAlchemy==1.3.0
text-unidecode==1.2
Unidecode==1.0.23
urllib3==1.24.1
uWSGI==2.0.18
validators==0.12.4
vine==1.2.0
Werkzeug==0.14.1
WTForms==2.2.1
WTForms-Components==0.10.3
